Updating the app entrypoint to support updated React Router. Refactoring disclaimer, footer and header out of main app. Removing model_name and eea_model (we just use the default model defined by backend). Removing SHOW_BRANDING and SHOW_LOGIN configuration values as they are never used.
